+++ src/services/RevertedBlockResolver.ts
@@ -12,7 +12,7 @@ class RevertedBlockResolver {
 
   async apply(lastSubmittedBlockId: number, nextBlockId: number, chainId: string): Promise<number> {
     if (lastSubmittedBlockId <= nextBlockId) {
-      this.logger.info(`DEBUG: [${chainId}] no reverts`);
+      this.logger.debug(`[${chainId}] no reverts`);
       return -1;
     }